make R.partition behave like its Haskell namesake
