+++ test/index.html
@@ -29,6 +29,7 @@
   <script src="test.range.js"></script>
   <script src="test.filter.js"></script>
   <script src="test.partial.js"></script>
+  <script src="test.groupBy.js"></script>
   <script src="test.partition.js"></script>
   <script src="test.objectBasics.js"></script>
   <script src="test.containsuniq.js"></script>
