+++ test/test.partition.js
@@ -2,48 +2,18 @@ var assert = require('assert');
 var Lib = require('./../ramda');
 
 describe('partition', function() {
-    var partition = Lib.partition, prop = Lib.prop;
+    var partition = Lib.partition;
 
-    it('should split the list into groups according to the partitioning function', function() {
-        var grade = function(score) {
-            return (score < 65) ? 'F' : (score < 70) ? 'D' : (score < 80) ? 'C' : (score < 90) ? 'B' : 'A';
-        };
-        var students = [
-            {name: 'Abby', score: 84},
-            {name: 'Brad', score: 73},
-            {name: 'Chris', score: 89},
-            {name: 'Dianne', score: 99},
-            {name: 'Eddy', score: 58},
-            {name: 'Fred', score: 67},
-            {name: 'Gillian', score: 91},
-            {name: 'Hannah', score: 78},
-            {name: 'Irene', score: 85},
-            {name: 'Jack', score: 69}
-        ];
-        var byGrade = function(student) {return grade(student.score || 0);};
-        assert.deepEqual(partition(byGrade, students), {
-            'A': [{name: 'Dianne', score: 99}, {name: 'Gillian', score: 91}],
-            'B': [{name: 'Abby', score: 84}, {name: 'Chris', score: 89}, {name: 'Irene', score: 85}],
-            'C': [{name: 'Brad', score: 73}, {name: 'Hannah', score: 78}],
-            'D': [{name: 'Fred', score: 67}, {name: 'Jack', score: 69}],
-            'F': [{name: 'Eddy', score: 58}]
-        });
+    it('splits a list into two lists according to a predicate', function() {
+        var pred = function(n) { return n % 2; };
+        assert.deepEqual(partition(pred, []), [[], []]);
+        assert.deepEqual(partition(pred, [0, 2, 4, 6]), [[], [0, 2, 4, 6]]);
+        assert.deepEqual(partition(pred, [1, 3, 5, 7]), [[1, 3, 5, 7], []]);
+        assert.deepEqual(partition(pred, [0, 1, 2, 3]), [[1, 3], [0, 2]]);
     });
-    
-    it('should be automatically curried', function() {
-        var splitByType = partition(prop("type"));
-        assert.deepEqual(splitByType([
-            {type: 'A', val: 10},
-            {type: 'B', val: 20},
-            {type: 'A', val: 30},
-            {type: 'A', val: 40},
-            {type: 'C', val: 50},
-            {type: 'B', val: 60}
-        ]), {
-            'A': [{type: 'A', val: 10}, {type: 'A', val: 30},{type: 'A', val: 40}],
-            'B': [{type: 'B', val: 20}, {type: 'B', val: 60}],
-            'C': [{type: 'C', val: 50}]
-        })
 
+    it('is curried', function() {
+        var polarize = partition(Boolean);
+        assert.deepEqual(polarize([true, 0, 1, null]), [[true, 1], [0, null]]);
     });
 });
